fix: corregir errores del panel admin y ampliar dashboard de Energy
Repara el TypeError de number_format en los widgets de KeyCounter (sum()
devolvía string en PostgreSQL), soluciona el thumbnail de imagen que no se
mostraba en Smart Plant Plants (disco incorrecto en ImageColumn), elimina la
página de mapa de AirFlight del backend (solo vive en el frontend) y sustituye
el widget único de Energy por un resumen completo (estado actual, hoy y 30
días) junto al gráfico de generación vs consumo a ancho completo.

+++ app/Filament/Admin/Widgets/Keyboard30DaysWidget.php
@@ -16,7 +16,7 @@ class Keyboard30DaysWidget extends BaseWidget
 
     protected function getStats(): array
     {
-        $pulsations = Keyboard::where('start_at', '>=', now()->subDays(30))->sum('pulsations') ?? 0;
+        $pulsations = (float) (Keyboard::where('start_at', '>=', now()->subDays(30))->sum('pulsations') ?? 0);
 
         return [
             Stat::make('Pulsaciones de Teclas (Últimos 30 días)', number_format($pulsations))
